requirements.txt
Flask==3.1.2
flask_limiter==3.12
psycopg2==2.9.11
python-dotenv==1.2.1
Werkzeug==3.1.5